weather cleaned up
