+++ .gitignore
@@ -12,3 +12,4 @@ web-report/
 
 # macOS
 .DS_Store
+ApiKeys.ts
