+++ utility/weatherConditions.ts
@@ -1,0 +1,51 @@
+export const weatherConditions = {
+  Rain: {
+    color: '#005BEA',
+    title: 'Raining',
+    subtitle: 'Get a cup of coffee',
+    icon: 'weather-rainy'
+  },
+  Clear: {
+    color: '#f7b733',
+    title: 'Sunny',
+    subtitle: 'It is hurting my eyes',
+    icon: 'weather-sunny'
+  },
+  Thunderstorm: {
+    color: '#616161',
+    title: 'A Storm is coming',
+    subtitle: 'Because Gods are angry',
+    icon: 'weather-lightning'
+  },
+  Clouds: {
+    color: '#1F1C2C',
+    title: 'Clouds',
+    subtitle: 'Everywhere',
+    icon: 'weather-cloudy'
+  },
+
+  Snow: {
+    color: '#00d2ff',
+    title: 'Snow',
+    subtitle: 'Get out and build a snowman for me',
+    icon: 'weather-snowy'
+  },
+  Drizzle: {
+    color: '#076585',
+    title: 'Drizzle',
+    subtitle: 'Partially raining...',
+    icon: 'weather-hail'
+  },
+  Haze: {
+    color: '#66A6FF',
+    title: 'Haze',
+    subtitle: 'Another name for Partial Raining',
+    icon: 'weather-hail'
+  },
+  Mist: {
+    color: '#3CD3AD',
+    title: 'Mist',
+    subtitle: "Don't roam in forests!",
+    icon: 'weather-fog'
+  }
+}
